ec2/cicd/in-place/stress-app/requirements.txt
click==8.1.3
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
py-healthcheck==1.10.1
six==1.16.0
typing-extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.1